+++ java/demo/src/main/java/com/google/demo/Results.java
@@ -16,7 +16,7 @@
  * @author Shaopeng Jia
  */
 
-package com.google.i18n.phonenumbers.demo;
+package com.google.demo;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.Locale.ENGLISH;
@@ -256,7 +256,7 @@ public class Results extends HttpServlet {
         .build();
     SoyTofu tofu = sfs.compileToTofu();
 
-    SoyTofu simpleTofu = tofu.forNamespace("demo.input");
+    SoyTofu simpleTofu = tofu.forNamespace("demo.output");
 
     return
         simpleTofu.newRenderer(
