--- java/demo/src/main/java/com/google/i18n/phonenumbers/demo/InputFormTemplates.java
@@ -1,95 +1,0 @@
-// This file was automatically generated by the Soy compiler.
-// Please don't edit this file by hand.
-// source: /Users/rogg/IdeaProjects/libphonenumber/java/demo/src/main/resources/com/google/i18n/phonenumbers.demo/input_form.soy
-
-package com.google.i18n.phonenumbers.demo;
-
-
-/**
- * Wrapper class containing {@link com.google.template.soy.data.SoyTemplate} builders for each
- * template in: input_form.soy.
- */
-@javax.annotation.Generated("com.google.template.soy.SoyParseInfoGenerator")
-public final class InputFormTemplates {
-
-  /** A map of filepath to symbol used for CSS resolution on server edit-refresh. */
-  private static final com.google.common.collect.ImmutableMap<java.lang.String, java.lang.String> __PROVIDED_CSS_MAP__ =   com.google.common.collect.ImmutableMap.<java.lang.String, java.lang.String>of();
-
-  /** A list of provided symbols used for css validation on edit refresh. */
-  private static final com.google.common.collect.ImmutableList<java.lang.String> __PROVIDED_CSS__ =   com.google.common.collect.ImmutableList.<java.lang.String>of();
-
-  /** Template params for demo.input.inputForm. */
-  public static final class InputForm extends com.google.template.soy.data.BaseSoyTemplateImpl {
-
-    private static final java.lang.String __NAME__ = "demo.input.inputForm";
-
-    /** Wraps a ListenableFuture<InputForm> as a SoyTemplate.AsyncWrapper<InputForm> */
-    public static com.google.template.soy.data.SoyTemplate.AsyncWrapper<InputForm> wrapFuture(com.google.common.util.concurrent.ListenableFuture<InputForm> paramsFuture) {
-      return new com.google.template.soy.data.SoyTemplate.AsyncWrapper<>(__NAME__, paramsFuture);
-    }
-
-    private InputForm(com.google.common.collect.ImmutableMap<java.lang.String, com.google.template.soy.data.SoyValueProvider> data) {
-      super(data);
-    }
-
-    @java.lang.Override
-    public final java.lang.String getTemplateName() {
-      return __NAME__;
-    }
-
-    /** Creates a new Builder instance. */
-    public static Builder builder() {
-      return new Builder();
-    }
-
-    /** {@param welcomeTitle} */
-    public static final com.google.template.soy.data.SoyTemplateParam<java.lang.String>
-        WELCOME_TITLE =
-            com.google.template.soy.data.SoyTemplateParam.standard(
-                "welcomeTitle",
-                /* required= */ true,
-                com.google.common.reflect.TypeToken.of(java.lang.String.class));
-
-    /** {@inject csp_style_nonce} Created by ContentSecurityPolicyNonceInjectionPass. */
-    public static final com.google.template.soy.data.SoyTemplateParam<java.lang.Object>
-        CSP_STYLE_NONCE =
-            com.google.template.soy.data.SoyTemplateParam.injected(
-                "csp_style_nonce",
-                /* required= */ false,
-                com.google.common.reflect.TypeToken.of(java.lang.Object.class));
-
-    private static final com.google.common.collect.ImmutableSet<com.google.template.soy.data.SoyTemplateParam<?>> __PARAMS__ = com.google.common.collect.ImmutableSet.of(
-        WELCOME_TITLE);
-
-    public static final class Builder extends com.google.template.soy.data.BaseSoyTemplateImpl.AbstractBuilder<Builder, InputForm> {
-
-      private Builder() {
-        super(1);
-      }
-
-      @java.lang.Override
-      protected com.google.common.collect.ImmutableSet<com.google.template.soy.data.SoyTemplateParam<?>> allParams() {
-        return __PARAMS__;
-      }
-
-      @java.lang.Override
-      protected InputForm buildInternal(com.google.common.collect.ImmutableMap<java.lang.String, com.google.template.soy.data.SoyValueProvider> data) {
-        return new InputForm(data);
-      }
-
-      /** Sets welcomeTitle. */
-      @com.google.errorprone.annotations.CanIgnoreReturnValue
-      public Builder setWelcomeTitle(java.lang.String value) {
-        return setParamInternal(WELCOME_TITLE, asString(value));
-      }
-
-      /** Future compatible version of {@link #setWelcomeTitle(java.lang.String)}. */
-      @com.google.errorprone.annotations.CanIgnoreReturnValue
-      public Builder setWelcomeTitleFuture(java.util.concurrent.Future<java.lang.String> future) {
-        return setParamInternal(WELCOME_TITLE, asFuture(future, AbstractBuilder::asString));
-      }
-
-    }
-  }
-
-}
