+++ java/demo/src/main/resources/com/google/demo/input.soy
