+++ java/demo/src/main/resources/com/google/demo/result.soy
