+++ java/demo/src/main/webapp/WEB-INF/web.xml
@@ -4,22 +4,22 @@
     xmlns="http://java.sun.com/xml/ns/javaee" xsi:schemaLocation="http://java.sun.com/xml/ns/javaee
   http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd">
   <servlet>
-    <servlet-class>com.google.i18n.phonenumbers.demo.Results</servlet-class>
+    <servlet-class>com.google.demo.Results</servlet-class>
     <servlet-name>PhoneNumberParser</servlet-name>
   </servlet>
   <servlet>
-    <servlet-class>com.google.i18n.phonenumbers.demo.InputForm</servlet-class>
-    <servlet-name>InputForm</servlet-name>
+    <servlet-class>com.google.demo.Input</servlet-class>
+    <servlet-name>Input</servlet-name>
   </servlet>
   <servlet-mapping>
     <servlet-name>PhoneNumberParser</servlet-name>
     <url-pattern>/phonenumberparser</url-pattern>
   </servlet-mapping>
   <servlet-mapping>
-    <servlet-name>InputForm</servlet-name>
+    <servlet-name>Input</servlet-name>
     <url-pattern>/inputform</url-pattern>
   </servlet-mapping>
   <welcome-file-list>
-    <welcome-file>InputForm</welcome-file>
+    <welcome-file>Input</welcome-file>
   </welcome-file-list>
 </web-app>
